fix(component): #501 Progress-old, create first UT
