+++ packages/element3/packages/progress-old/index.js
@@ -1,0 +1,7 @@
+import ElProgress from './Progress.vue'
+
+ElProgress.install = function (app) {
+  app.component(ElProgress.name, ElProgress)
+}
+
+export default ElProgress
